add dialogs and move features between activity/fragment fixed features add all img from device
