+++ app/src/main/java/com/example/gallery/ui/profile/ProfileViewModel.java
@@ -140,195 +140,6 @@ public class ProfileViewModel extends BaseViewModel<ProfileNavigator> {
 
 
 
-    public void addImageFromLink() {
-
-        getNavigator().getmProfileBinding().txtDownloadStatus.setText("Downloaded Status: downloading");
-        getNavigator().getmProfileBinding().txtDownloadStatus.setTextColor(Color.parseColor("#808080"));
-
-        String imageUrl = getNavigator().getmProfileBinding().editTextImageUrl.getText().toString();
-
-        //  System.out.println("Image Url: " + imageUrl);
-        if (!isValidUrl(imageUrl)) {
-            getNavigator().getmProfileBinding().txtDownloadStatus.setText("Error: Invalid URL");
-            getNavigator().getmProfileBinding().txtDownloadStatus.setTextColor(Color.parseColor("#FF0000"));
-
-            return;
-        }
-
-
-        OkHttpClient client = new OkHttpClient();
-
-        Request request = new Request.Builder()
-                .url(imageUrl)
-                .build();
-
-        client.newCall(request).enqueue(new Callback() {
-            @Override
-            public void onFailure(Call call, IOException e) {
-                e.printStackTrace();
-                // Handle OkHttp failure...
-                //  System.out.println("Have something wrong ");
-                getNavigator().getmProfileBinding().txtDownloadStatus.setText("Downloaded Status: failed");
-                getNavigator().getmProfileBinding().txtDownloadStatus.setTextColor(Color.parseColor("#FF0000"));
-
-
-            }
-
-            @Override
-            public void onResponse(Call call, Response response) throws IOException {
-                if (response.isSuccessful()) {
-                    // Get the file extension from the Content-Type header
-                    String contentType = response.header("Content-Type");
-                    String fileExtension = getFileExtension(contentType);
-                    if (!isImageContentType(contentType)) {
-                        getNavigator().getmProfileBinding().txtDownloadStatus.setText("Error: Invalid image URL");
-                        getNavigator().getmProfileBinding().txtDownloadStatus.setTextColor(Color.parseColor("#FF0000"));
-
-                        return;
-                    }
-
-                    //  System.out.println("Content Type: " + contentType);
-
-                    // Get the file name from the URL
-                    String fileName = addRandomNumberToFileName(getFileNameFromUrl(imageUrl)) + fileExtension;
-                    //  System.out.println("File name from url: " + fileName);
-
-                    // Save the image to a file
-                    File appDirectory = new File(App.getInstance().getExternalFilesDir(null), "Images/FromUrls");
-                    if (!appDirectory.exists()) {
-                        appDirectory.mkdirs();
-                    }
-
-                    File imageFile = new File(appDirectory, fileName);
-
-                    try (OutputStream stream = new FileOutputStream(imageFile)) {
-                        stream.write(response.body().bytes());
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-
-                    // Get image path
-                    String imgPath = imageFile.getAbsolutePath();
-                    //  System.out.println("Image path: " + imgPath);
-
-
-                    // Save media item info to database
-                    MediaItem item = new MediaItem();
-                    item.setPath(imgPath);
-                    item.setOrigin(imageUrl);
-                    item.setCreationDate(Date.parse(new Date().toString()));
-//                    item.setId(new Random().nextInt(10000));
-                    item.setUserID(getDataManager().getCurrentUserId());
-                    item.setAlbumName("Download");
-
-                    // check if album not exist in table albums
-
-                    MediaItemViewModel.getInstance().insert(item);
-//                    Toast.makeText(App.getInstance(), "Insert successfully", Toast.LENGTH_SHORT).show();
-                    //  System.out.println("Insert media item success from profile view model");
-
-
-
-                    // Get the number of rows in the database
-//                    //  System.out.println("Number of media items after download image: " + MediaItemViewModel.getInstance().getNumberOfMediaItems().getValue());
-
-                    getNavigator().getmProfileBinding().txtDownloadStatus.setText("Status: download a " + fileExtension +" successfully!");
-                    getNavigator().getmProfileBinding().txtDownloadStatus.setTextColor(Color.parseColor("#008000"));
-                }
-            }
-        });
-
-        // display image
-        Glide.with(App.getInstance())
-                .asBitmap()
-                .load(imageUrl)
-                .into(getNavigator().getmProfileBinding().imageViewDownloaded);
-        getNavigator().getmProfileBinding().imageViewDownloaded.setVisibility(View.VISIBLE);
-
-
-    }
-
-    public void displayImageDownloaded(String imgPath) {
-        Bitmap bitmap = BitmapFactory.decodeFile(imgPath);
-
-        // Set the bitmap to the ImageView
-        ImageView imageView = getNavigator().getmProfileBinding().imageViewDownloaded;
-        imageView.setImageBitmap(bitmap);
-        getNavigator().getmProfileBinding().imageViewDownloaded.setVisibility(View.VISIBLE);
-
-
-    }
-
-
-    private boolean isImageContentType(String contentType) {
-        // Check if the content type corresponds to an image type
-        // You can customize this method based on the specific image types you want to support
-        return contentType != null && contentType.startsWith("image/");
-    }
-
-    private boolean isValidUrl(String urlString) {
-        try {
-            new URL(urlString);
-            return true;
-        } catch (MalformedURLException e) {
-            return false;
-        }
-    }
-
-    private static String getFileExtension(String contentType) {
-        if (contentType != null && contentType.contains("/")) {
-            return "." + contentType.substring(contentType.lastIndexOf("/") + 1);
-        }
-        return "";
-    }
-
-    private void fetchImageContentType(String imageUrl) {
-        OkHttpClient client = new OkHttpClient();
-
-        Request request = new Request.Builder()
-                .url(imageUrl)
-                .head() // Use a HEAD request to fetch headers only
-                .build();
-
-        client.newCall(request).enqueue(new Callback() {
-            @Override
-            public void onFailure(Call call, IOException e) {
-                e.printStackTrace();
-            }
-
-            @Override
-            public void onResponse(Call call, Response response) {
-                // Get content type from the response headers
-                String contentType = response.header("Content-Type");
-                //  System.out.println("Content Type: " + contentType);
-                // Handle the content type as needed
-            }
-        });
-    }
-
-    public static String getFileNameFromUrl(String url) {
-        if (url == null || url.isEmpty()) {
-            return null;
-        }
-
-        // Use URI to parse the URL
-        Uri uri = Uri.parse(url);
-
-        // Get the last path segment, which is usually the filename
-        String lastPathSegment = uri.getLastPathSegment();
-
-        // You can perform additional checks or modifications if needed
-        return lastPathSegment;
-    }
-
-    private static String addRandomNumberToFileName(String fileName) {
-        Random random = new Random();
-        int randomNumber = random.nextInt(1000); // Adjust the range as needed
-
-        // Append the random number to the filename
-        return randomNumber + "_" + fileName;
-    }
-
 
     public void backup() {
         // TODO: backup all user data to local storage
@@ -507,15 +318,6 @@ public class ProfileViewModel extends BaseViewModel<ProfileNavigator> {
     }
 
 
-
-
-// Để chạy task:
-
-    public void clearText() {
-
-        getNavigator().getmProfileBinding().editTextImageUrl.setText("");
-        getNavigator().getmProfileBinding().imageViewDownloaded.setVisibility(View.GONE);
-    }
 
 
 }
