+++ app/src/main/res/layout/download_img_url.xml
@@ -3,31 +3,84 @@
               android:layout_width="match_parent"
               android:layout_height="wrap_content"
               android:orientation="vertical"
-              android:padding="16dp">
+              android:padding="20dp"
+    android:background="#F0FFFF"
+    >
+
+    <ImageView
+        android:layout_width="200dp"
+        android:layout_height="200dp"
+        android:layout_marginTop="6dp"
+        android:id="@+id/imgVDownloaded"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="6dp"
+        android:src="@drawable/ic_launcher_background"
+        android:layout_gravity="center"
+        android:visibility="gone"
+        />
 
     <TextView
-        android:id="@+id/tvStatus"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Status: "
-        android:textStyle="bold"
-        android:textSize="18sp" />
+        android:hint="Download Status: "
+        android:textSize="16dp"
+        android:layout_marginTop="20dp"
+        android:id="@+id/tvStatus"
+        android:textColor="@color/teal_700"
+        />
 
     <EditText
         android:id="@+id/etUrl"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="Enter URL" />
 
-    <Button
-        android:id="@+id/btnDownload"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Download" />
+        android:hint="Enter Image URL"
+        android:layout_marginTop="20dp"
+        android:layout_marginBottom="20dp"
+        android:inputType="text"/>
 
-    <Button
-        android:id="@+id/btnClear"
+    <LinearLayout
+        android:id="@+id/linearLayoutAddFromLink"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Clear" />
+        android:orientation="horizontal"
+        android:layout_marginBottom="16dp"
+        >
+        <com.google.android.material.button.MaterialButton
+            style="?attr/materialButtonOutlinedStyle"
+
+            android:id="@+id/btnDownload"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:text="Download"
+            android:layout_marginTop="6dp"
+            android:layout_weight="1"
+            android:backgroundTint="#C6E2FF"
+
+            />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            />
+
+        <com.google.android.material.button.MaterialButton
+            style="?attr/materialButtonOutlinedStyle"
+
+            android:id="@+id/btnClear"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_marginTop="6dp"
+            android:layout_weight="1"
+            android:backgroundTint="#C6E2FF"
+
+
+            android:text="Clear "
+
+            />
+
+    </LinearLayout>
+
+
 </LinearLayout>
